+++ src/screens/onboarding/PhoneVerificationScreen.tsx
@@ -3,7 +3,7 @@ import React, { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { View } from 'react-native';
 import { useOnboardingStore } from '@/store/onboarding';
-import { useTimer } from '@/hooks/useTimer';
+import { useTimer } from '@/hooks';
 import { saveTokens } from '@/utils/service/auth';
 import { logError } from '@/utils/service/error';
 import { formatPhone } from '@/utils/service/phone';
